requirements.txt
netCDF4==1.6.5
numpy==1.26.2
pandas==2.1.3
pymongo==3.12.0
python-dotenv==1.0.1
tqdm
